Correcion de Proxy Modulacion
