+++ netback-proxy/.env.example
@@ -1,0 +1,42 @@
+# ============================================================================
+# Configuración de Django API
+# ============================================================================
+DJANGO_API_PROTOCOL=http
+DJANGO_API_URL=localhost
+DJANGO_API_PORT=8000
+
+# ============================================================================
+# Configuración de FastAPI
+# ============================================================================
+FASTAPI_HOST=0.0.0.0
+FASTAPI_PORT=8080
+FASTAPI_DEBUG=false
+
+# ============================================================================
+# Configuración de CORS (usar strings separados por comas)
+# ============================================================================
+ALLOW_ORIGINS=http://localhost:3000,http://localhost:8080,http://127.0.0.1:3000
+ALLOW_CREDENTIALS=true
+ALLOW_METHODS=GET,POST,PUT,PATCH,DELETE,OPTIONS
+ALLOW_HEADERS=Content-Type,Authorization,X-Requested-With
+
+# ============================================================================
+# Configuración de Timeouts
+# ============================================================================
+DJANGO_REQUEST_TIMEOUT=30
+
+# ============================================================================
+# Configuración de Logging
+# ============================================================================
+LOG_LEVEL=INFO
+
+# ============================================================================
+# Configuración de Seguridad
+# ============================================================================
+SECRET_KEY=your-super-secret-key-here-change-in-production
+
+# ============================================================================
+# Configuración para Producción (opcional)
+# ============================================================================
+# SENTRY_DSN=https://your-sentry-dsn-here
+# ENVIRONMENT=production
